Succumb to those pesky trends
After field tests in an actual extension I've come to the conclusion that it feels odd not to have pointer cursor for trendy flat buttons. Traditionally buttons use the default cursor as the button itself communicates clearly it is a clickable element. However a recent trend, seemingly from web developers not knowing pointer traditionally specifically means link, not just clickable in general, the use of pointer cursors for buttons have become very widespread.

It is also worth mentioning that the Extension Manager, which this library is graphically based on, uses pointer cursors.